MainActivity.kt is now getting remote data from OpenWeatherApi.kt

+++ app/src/main/java/com/example/carrentalfinder/data/models/Main.kt
@@ -4,10 +4,10 @@ package com.example.carrentalfinder.data.models
  * Data class for main entity
  */
 data class Main (
-    val temp: Int,
-    val feels_like: Int,
-    val temp_min: Int,
-    val temp_max: Int,
+    val temp: Double,
+    val feels_like: Double,
+    val temp_min: Double,
+    val temp_max: Double,
     val pressure: Int,
     val humidity: Int
     )
